tabletserver: strictMode check moved to QueryEngine
It makes more sense for strict mode check to be moved to QueryEngine
because it's the one that depends on that property.
